risk-api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.2.1
distro==1.9.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.0
lightgbm==4.6.0
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
